fixed api categories and products
