+++ server/api/products.spec.js
@@ -6,7 +6,7 @@ const db = require('../db')
 const app = require('../index')
 const Product = db.model('product')
 const Review = db.model('review')
-const Category = db.model('category');
+
 
 describe('Product routes', () => {
   beforeEach(() => {
@@ -19,7 +19,7 @@ describe('Product routes', () => {
       name: 'El Flamo',
       description: 'hotter than hell, (batteries not included)',
       price: 40.44,
-      inventory: 20
+      inventory: 20,
     }
 
     const newReview = {
@@ -29,8 +29,8 @@ describe('Product routes', () => {
       productId: 1
     }
 
-    beforeEach(() => {
-      return Product.create(newProduct)
+    beforeEach( async () => {
+      await Product.create(newProduct)
     })
 
     beforeEach(() => {
@@ -87,30 +87,7 @@ describe('Product routes', () => {
         });
     })
   })
-  describe('/api/products/categories', () => {
-    const cat1 = { name: 'analog' };
-    const cat2 = { name: 'gas-based' };
-    const cat3 = { name: 'real-flame' };
-    beforeEach( () => {
-      return Category.create(cat1);
-    })
-    beforeEach( () => {
-      return Category.create(cat2);
-    })
-    beforeEach( () => {
-      return Category.create(cat3);
-    })
 
-    it('GET /api/products/categories', () => {
-      return request(app)
-        .get('/api/products/categories')
-        .expect(200)
-        .then(res => {
-          expect(res.body).to.be.an('array')
-          expect(res.body[0].name).to.be.equal('analog')
-        })
-    })
-  })
 
 
 })
